feat(simulator): ASID-keyed VMEM shared-overlay + 16-core sharer_mask + observability
The 8-core canneal V2 run produced a sharer-count histogram with bin[1] = 100.000%
over 19.9 M evictions. Root-caused to VMEM's per-CPU isolation: vpage_to_ppage_map
keys on (cpu_num, vaddr), so two cores' identical VAs deterministically map to
different physical pages — no LLC line can ever be touched by more than one core.
This made the +20×sharers bias, the sharer-count perceptron feature, and the
A.1/A.2 invalidation hook empirically inert on the existing PARSEC traces.

This change adds three orthogonal pieces:

1. ASID-keyed VMEM shared-overlay (vmem.h, vmem.cc + mirror to sibling):
   - New SHARED_ASID = UINT32_MAX sentinel.
   - cpu_to_shared_asid map (per-CPU effective ASID).
   - va_to_pa now keys on cpu_to_shared_asid[cpu_num] when present, else cpu_num
     (default behavior preserved when no shared CPUs configured).
   - set_shared_cpus(vec) marks each listed CPU as sharing SHARED_ASID; idempotent
     and additive.
   - Static aliased_fills counter — incremented when a request for a shared vpage
     finds it already mapped by a DIFFERENT core. Printed at end of sim by main.cc.
   - PTEs deliberately NOT shared (get_pte_pa unchanged); page tables are per-process.

2. Config plumbing (parse.py, instantiation_file.py, btp_config.json,
   btp_8core_config.json):
   - New optional JSON field virtual_memory.vmem_shared_cpus (default []).
   - instantiation_file emits vmem.set_shared_cpus({...}) in the generated env
     constructor body when the list is non-empty.

3. Observability (cache_stats.h + .cc + plain_printer.cc + cache.cc):
   - New coherence_write_hit_other_sharer_events counter (distinct from existing
     coherence_invalidations which counts bits cleared; this counts events).
   - Sharer histogram already copied to roi_stats in end_phase.
   - Both new counters printed by plain_printer per cache.

Also includes the prior session's 16-core sharer_mask widening (uint8_t -> uint16_t)
in block.h, cache.cc, and coalesce.cc, completing 16-core scaling support.

Verified locally:
  - Default config (no shared CPUs): VMEM ALIASED FILLS = 0,
    LLC COHERENCE INVALIDATIONS = 0, LLC COHERENCE EVENTS = 0.
    Regression test passes — existing semantics preserved byte-for-byte.
  - Shared config (CPUs 0..3 + canneal 50 k smoke): VMEM ALIASED FILLS = 30.
    Mechanism fires end-to-end. LLC sharer-mask propagation requires a workload
    where the aliased pages actually reach the LLC (see bench/synth_coherence.c).

Mirrored to sibling /home/rajharsh/programming-playground/repos/ChampSim for the
build-tree gotcha (OPEN_DECISIONS #16): inc/vmem.h, inc/cache_stats.h.

Per the plan at ~/.claude/plans/ok-analyzew-the-entire-crystalline-reef.md.

+++ simulator/inc/block.h
@@ -22,7 +22,7 @@ struct cache_block {
 
   // COALESCE Coherence Extensions
   MESI_State state = INVALID;
-  uint8_t sharer_mask = 0;
+  uint16_t sharer_mask = 0;  // 16-bit to support up to 16 cores (scaling study)
 };
 } // namespace champsim
